requirements.txt
Django==2.2.3
djangorestframework==3.10.3
Markdown==3.1.1
django-cors-headers==3.7.0